pyproject.toml
[project]
name = "nanodqn"
version = "0.1.0"
description = "Reimplementation of CleanRL DQN in JAX+Equinox and jaxtyping"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "wandb",
    "flax",
    "jax[cuda12]", # nvidia support by default
    "optax",
    "gymnasium[classic-control]<1",
    "numpy<2", # cuz stable baselines3 die otherwise
    "tyro",
    "tqdm",
    "moviepy", # videos
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "equinox>=0.12.2",
    "jaxtyping>=0.3.2",
    "gymnax>=0.0.9",
]


[tool.ruff]
line-length = 100

[tool.uv.sources]
gymnax = { path = "gymnax", editable = true }
